feat: :sparkles: enhance user deletion process by adding password verification before deletion

+++ src/app/modules/auth/auth.service.ts
@@ -331,10 +331,21 @@ const socialLoginFromDB = async (payload: IUser) => {
 }
 
 // delete user
-const deleteUserFromDB = async (user: JwtPayload) => {
+// delete user
+const deleteUserFromDB = async (user: JwtPayload, password: string) => {
 
-    const isExistUser = await User.findByIdAndDelete(user.id);
+    const isExistUser = await User.findById(user.id).select('+password');
     if (!isExistUser) {
+        throw new ApiError(StatusCodes.BAD_REQUEST, "User doesn't exist!");
+    }
+
+    //check match password
+    if (password && !(await User.isMatchPassword(password, isExistUser.password))) {
+        throw new ApiError(StatusCodes.BAD_REQUEST, 'Password is incorrect');
+    }
+
+    const updateUser = await User.findByIdAndDelete(user.id);
+    if (!updateUser) {
         throw new ApiError(StatusCodes.BAD_REQUEST, "User doesn't exist!");
     }
     return;
